Bug fixes + Application code for I2C Master Data sending , to be tested

+++ stm32f4xx_drivers/drivers/Inc/stm32f4xx_i2c_driver.h
@@ -15,9 +15,9 @@
 typedef struct
 {
 	uint32_t I2C_SCLSpeed;
-	uint8_t  I2C_DeviceAddress;
-	uint8_t  I2C_AckControl;
-	uint8_t  I2C_FMDutyCycle;
+	uint32_t  I2C_DeviceAddress;
+	uint32_t  I2C_AckControl;
+	uint32_t  I2C_FMDutyCycle;
 
 }I2C_Config_t;
